Created and Implemented Permissions: IsAuthor
IsAuthor class is responsible for checking if the request author matches the original author. This ensures only authorized users has permissions to modify any data.

IsAuthenticated ensures that the user has valid token or authentication before doing any actions.

+++ apps/posts/views.py
@@ -3,8 +3,10 @@ from django.core.exceptions import ValidationError
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
+from rest_framework.permissions import IsAuthenticated
 from .models import Post, Comment
 from .serializers import PostSerializer, CommentSerializer
+from .permissions import IsAuthor
 
 # Post List View (GET all posts)
 class PostListView(APIView):
@@ -33,14 +35,18 @@ class PostListView(APIView):
 # Post Detail View (GET, PUT, DELETE a single post)
 class PostDetailView(APIView):
     serializer_class = PostSerializer
+    permission_classes = [IsAuthenticated, IsAuthor]
 
     def get(self, request, pk):
         try:
             # Get specific post based on primary key
             post = Post.objects.get(pk=pk)
+            # Check authentication and permissions
+            self.check_object_permissions(request, post)
             # Serialize the post object
             serializer = PostSerializer(post)
             # Return the serialized JSON to http response
+                        # Return the serialized JSON to http response
             return Response(serializer.data)
         except Post.DoesNotExist:
             # Throw an error if the post with the specified id is not found
@@ -60,6 +66,8 @@ class PostDetailView(APIView):
         try:
             # Check if the serializer is valid, raise exception if the request data contains invalid field
             if serializer.is_valid(raise_exception=True):
+                # Check authentication and permissions
+                self.check_object_permissions(request, post)
                 # Save the serializer data to the db
                 serializer.save()
                 # Return an ok response
@@ -76,6 +84,8 @@ class PostDetailView(APIView):
         try:
             # Get specified post based on primary key
             post = Post.objects.get(pk=pk)
+            # Check authentication and permissions
+            self.check_object_permissions(request, post)
             # Delete the post from the db
             post.delete()
             # Return an ok response
@@ -130,6 +140,7 @@ class CommentListView(APIView):
 # Comment Detail View (GET, PUT, DELETE a specific comment)
 class CommentDetailView(APIView):
     serializer_class = CommentSerializer
+    permission_classes = [IsAuthenticated, IsAuthor]
 
     def get(self, request, post_id, comment_id):
         try:
@@ -160,9 +171,13 @@ class CommentDetailView(APIView):
             comment = Comment.objects.get(pk=comment_id, post=post)
         except Comment.DoesNotExist:
             return Response({'Message': 'Comment not found'}, status=status.HTTP_404_NOT_FOUND)
+        
+        # Check authentication and permissions
+        self.check_object_permissions(request, comment)
 
         # Deserialize and update the comment
         serializer = CommentSerializer(comment, data=request.data)
+        
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
@@ -182,6 +197,9 @@ class CommentDetailView(APIView):
             comment = Comment.objects.get(pk=comment_id, post=post)
         except Comment.DoesNotExist:
             return Response({'Message': 'Comment not found'}, status=status.HTTP_404_NOT_FOUND)
+        
+        # Check authentication and permissions
+        self.check_object_permissions(request, comment)
 
         # Delete the comment
         comment.delete()
